Refactoring installing script

+++ scripts/pvem_install.sh
@@ -20,7 +20,7 @@ _pvem_install() {
     fi
 
     # Search the python FTP server for the latest version that starts with the
-    # given version
+    # given version (3.11 -> 3.11.8, 3.11.7 -> 3.11.7...)
     version=$(curl -s https://www.python.org/ftp/python/ | grep -oE "[0-9]+\.[0-9]+\.[0-9]+" | grep -E "^$target_version(?:\D|$)" | sort -V | tail -n 1)
 
     if [ -z "$version" ]; then
@@ -45,6 +45,7 @@ _pvem_install() {
         return 1
     fi
 
+    printf "\n"
     if ! __pvem_download_and_install_version "$target_version"; then
         printf "${C_RED}Error: Python version $target_version could not be installed\n"
         return 1
@@ -60,19 +61,50 @@ _pvem_install() {
 #   $1: Python version to install
 # Return: 0 if the python version was installed, 1 otherwise
 __pvem_download_and_install_version() {
-    target_version=$1
+    if [ -d "$VERSIONPATH/tmp" ]; then
+        rm -rf "$VERSIONPATH/tmp"
+    fi
+
+    VERSION=$1
+    UNPACK_PATH="$VERSIONPATH/tmp"
+    TAR_PATH="$UNPACK_PATH/Python-$VERSION.tgz"
+    INSTALL_PATH="$VERSIONPATH/$VERSION"
+    
+    if ! __pvem_download_python_source "$VERSION" "$TAR_PATH"; then
+        return 1
+    fi
+
+    if ! __pvem_unpack_python_source "$TAR_PATH" "$UNPACK_PATH"; then
+        return 1
+    fi
+
+    install_exit_status=0
+    if ! __pvem_install_python_source "$UNPACK_PATH/Python-$VERSION" "$INSTALL_PATH"; then
+        install_exit_status=1
+    fi
 
     if [ -d "$VERSIONPATH/tmp" ]; then
         rm -rf "$VERSIONPATH/tmp"
     fi
 
-    mkdir -p "$VERSIONPATH/tmp"
-    
-    printf "\n"
-    printf "${C_BLUE}Downloading Python version $target_version${C_RESET} "
+    return $install_exit_status
+}
+
+# Function: __pvem_download_python_source
+# Summary: Download the source code of a python version
+# Parameters:
+#  $1: Python version to download
+#  $2: Target path for the tar source code
+# Return: 0 if the source code was downloaded, 1 otherwise
+__pvem_download_python_source() {
+    version=$1
+    target_path=$2
 
     WGET_LOG_FILE=$(mktemp)
-    wget -O "$VERSIONPATH/tmp/Python-$target_version.tgz" "https://www.python.org/ftp/python/$target_version/Python-$target_version.tgz" 2>$WGET_LOG_FILE
+
+    printf "${C_BLUE}Downloading source code${C_RESET} "
+    mkdir -p $(dirname $target_path)
+    wget -O $target_path "https://www.python.org/ftp/python/$version/Python-$version.tgz" 2>$WGET_LOG_FILE
 
     if [ $? -ne 0 ]; then
         printf "\n"
@@ -80,13 +112,28 @@ __pvem_download_and_install_version() {
         return 1
     fi
 
-    rm $WGET_LOG_FILE 2>/dev/null
-    printf "Done.\n"
+    if [ -f $WGET_LOG_FILE ]; then
+        rm $WGET_LOG_FILE
+    fi
 
-    printf "${C_BLUE}Extracting Python version $target_version${C_RESET} "
+    printf "Done.\n"
+    return 0
+}
+
+# Function: __pvem_unpack_python_source
+# Summary: Extract the source code of a python version
+# Parameters:
+#   $1: Path to the tar source code
+#   $2: Path to extract the source code
+# Return: 0 if the source code was extracted, 1 otherwise
+__pvem_unpack_python_source() {
+    source_path=$1
+    target_path=$2
 
     TAR_LOG_FILE=$(mktemp)
-    tar -zxf "$VERSIONPATH/tmp/Python-$target_version.tgz" -C "$VERSIONPATH/tmp" 2>$TAR_LOG_FILE
+
+    printf "${C_BLUE}Extracting source code${C_RESET} "
+    tar -zxf $source_path -C $target_path 2>$TAR_LOG_FILE
 
     if [ $? -ne 0 ]; then
         printf "\n"
@@ -94,16 +141,32 @@ __pvem_download_and_install_version() {
         return 1
     fi
 
-    rm $TAR_LOG_FILE 2>/dev/null
-    printf "Done.\n"
+    if [ -f $TAR_LOG_FILE ]; then
+        rm $TAR_LOG_FILE
+    fi
 
-    printf "${C_BLUE}Installing Python version $target_version${C_RESET}\n"
+    printf "Done.\n"
+    return 0
+}
+
+# Function: __pvem_install_python_source
+# Summary: Install the source code of a python version
+# Parameters:
+#  $1: Path to the source code
+#  $2: Path to install the source code
+# Return: 0 if the source code was installed, 1 otherwise
+__pvem_install_python_source() {
+    source_path=$1
+    target_path=$2
+
     INSTALL_LOG_FILE=$(mktemp)
     INSTALL_EXIT_STATUS_FILE=$(mktemp)
+
+    printf "${C_BLUE}Installing Python${C_RESET}\n"
     (
         set -e
-        cd "$VERSIONPATH/tmp/Python-$target_version" &&
-        ./configure --prefix="$VERSIONPATH/$target_version" &&
+        cd $source_path &&
+        ./configure --prefix=$target_path &&
         make -j4 &&
         make install
         echo $? > $INSTALL_EXIT_STATUS_FILE
@@ -116,15 +179,14 @@ __pvem_download_and_install_version() {
         exit_status=0
     fi
 
-    if [ -d "$VERSIONPATH/tmp" ]; then
-        rm -rf "$VERSIONPATH/tmp"
-    fi
-
     if [ $exit_status -ne 0 ]; then
         tail -n 10 $INSTALL_LOG_FILE
-        rm -rf "$VERSIONPATH/$target_version"
+        rm -rf $target_path
     fi
 
-    rm $INSTALL_LOG_FILE 2>/dev/null
+    if [ -f $INSTALL_LOG_FILE ]; then
+        rm $INSTALL_LOG_FILE
+    fi
+
     return $exit_status
 }
